requirements.txt
streamlit
langdetect
deep-translator
sentence-transformers
PyPDF2
faiss-cpu
